requirements.txt
Django==2.2
djangorestframework==3.13.1
markdown==3.3.6
pillow==9.1.0
requests==2.27.1
django-filter
django-guardian
coreapi
django-cors-headers